move SavingsGoalData into SharedSavingsGoalManager
